// frontend/src/app/models/user.ts
export default class User {
    _id: string = '';
    korisnickoIme: string = '';
    lozinka: string = '';
    ime: string = '';
    prezime: string = '';
    adresa: string = '';
    pol: string = '';
    telefon: string = '';
    email: string = '';
    profilnaSlika: string = '';
    tipKorisnika: string = '';
    brojKartice: string = '';
    status: string = '';
    firma: string = '';
    checked: boolean = false;
}